v2 reorganize directory structure.
