+++ packages/core/src/core.ts
@@ -1,9 +1,15 @@
+import { AccessorComponentType, AccessorComponentTypeData, AccessorType, AccessorTypeData, BufferViewTarget } from './constants';
+import { Logger, LoggerVerbosity } from './logger';
+import { NodeIO, WebIO } from './io';
+
+import { Element } from './elements/element';
 import { GLTFContainer } from './container';
 import { GLTFUtil } from './util';
-import { NodeIO, WebIO } from './io';
-import { Logger, LoggerVerbosity } from './logger';
-import { AccessorType, AccessorTypeData, AccessorComponentType, AccessorComponentTypeData, BufferViewTarget } from './constants';
-import { Graph, Root, Node, Mesh, Material, Element } from './v2/elements';
+import { Graph } from './graph/graph';
+import { Material } from './elements/material';
+import { Mesh } from './elements/mesh';
+import { Node } from './elements/node';
+import { Root } from './elements/root';
 
 export {
     GLTFUtil, GLTFContainer,
