+++ packages/core/src/math.ts
